Selenium_Requests_test/requirements.txt
selenium>=4.9.0
webdriver-manager>=3.8.6
pytest>=7.3.1
pytest-html>=3.2.0
allure-pytest>=2.13.2
requests>=2.29.0
pandas>=2.0.0
